refactor iconForDomain function
